Reworked Mouse TouchPad implementation (#449)
And analogous improvements to Mouse Joystick implementation.

This fixes #417 and #420

Two biggest changes:
- Moved from MoveTo() to MoveBy() implementation. Not because I prefer one or
  another but MoveTo() doesn't work with FPP games that center the cursor. Not
  sure why and did not manage to make it work despite several tries. The
  drawback of MoveBy() is that it's int only and small touchpad movements can be
  subpixel, so I've added a code that mitigates that and caches the float
  movement that has been left unused after a move cycle. This way very slow
  touchpad movement still progresses properly.
  Applied this to Joystick as well which visibly improves it even more.
- I've moved away from the hack to detect the touchpad touch to properly
  detecting the touch using device's sensors. Unfortunately the implementation
  is another hack that works only because the whole system is not event
  based. It might need a rework in the future.

Smaller changes:
- fixed deadzone application for joystick (cast to float before dividing)
- less vectors
- cleaner and more readable code (akin to joystick)
- some consistency changes to joystick as well
- common code for joystick/touchpad
- invert axis is applied for touchpad now
- removed deadzone definitions for touchpads as they make no sense
- changed MouseSimulator.MoveTo() to double as the underlying
  InputSimulator.Mouse.MoveMouseTo() is double but this function is unused now
  so can probably be removed altogether
- sensitivity changed to a safe (somewhat slow) value of 25 akin to joystick
  (both in the templates and in the MouseActions)

+++ HandheldCompanion/Simulators/MouseSimulator.cs
@@ -79,7 +79,8 @@ namespace HandheldCompanion.Simulators
             }
         }
 
-        public static void MoveTo(int x, int y)
+        // TODO: unused, remove?
+        public static void MoveTo(double x, double y)
         {
             try
             {
